requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.0
pillow==10.0.0
opencv-python==4.8.0.74
scikit-learn==1.3.0
gradio==3.50.0
requests==2.31.0
numpy==1.24.3